+++ tests/automl/batch_generator_test.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # flake8: noqa
-"""Tests `nineturn.dtdg.models.decoder.torch.sequentialDecoder.implicitTimeModels` package."""
+"""Tests `spiro.dtdg.models.decoder.torch.sequentialDecoder.implicitTimeModels` package."""
 import tensorflow as tf
 import numpy as np
 from tests.core.common_functions import *
@@ -13,12 +13,12 @@ times = np.array([2006, 2007, 2008,2009])
 def test_batch_generator():
     """Test nodebatchgenerator"""
     clear_background()
-    from nineturn.core.config import set_backend
-    from nineturn.core.backends import TENSORFLOW
+    from spiro.core.config import set_backend
+    from spiro.core.backends import TENSORFLOW
     set_backend(TENSORFLOW)
-    from nineturn.dtdg.types import CitationGraph
-    from nineturn.automl.tf.prepare_dataset import prepare_citation_task, TARGET, LABEL
-    from nineturn.automl.tf.batch_training import NodeBatchGenerator
+    from spiro.dtdg.types import CitationGraph
+    from spiro.automl.tf.prepare_dataset import prepare_citation_task, TARGET, LABEL
+    from spiro.automl.tf.batch_training import NodeBatchGenerator
     this_graph = CitationGraph(edges, nodes, times)
     prepare_citation_task(this_graph, validating_snapshots=1,minimum_citation=0)
     gen = NodeBatchGenerator(this_graph, 2,1,3,False)
