+++ tests/core/layers/tf/test_time2vec.py
@@ -1,23 +1,23 @@
 #!/usr/bin/env python
 # flake8: noqa
-"""Tests `nineturn.core.config` package."""
+"""Tests `spiro.core.config` package."""
 
 import os
 import numpy as np
 import tensorflow as tf
-from nineturn.core.backends import TENSORFLOW, PYTORCH
+from spiro.core.backends import TENSORFLOW, PYTORCH
 from tests.core.common_functions import *
-from nineturn.core.logger import get_logger
+from spiro.core.logger import get_logger
 
 logger = get_logger()
 
 
 def test_time2vec_tf():
     clear_background()
-    from nineturn.core.config import _BACKEND, set_backend
+    from spiro.core.config import _BACKEND, set_backend
 
     set_backend(TENSORFLOW)
-    from nineturn.core.layers import Time2Vec
+    from spiro.core.layers import Time2Vec
     
     features = 5
     kernel = 4
