+++ tests/dtdg/test_dataloader_torch.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # flake8: noqa
-"""Tests `nineturn.dtdg.dataloader` package."""
+"""Tests `spiro.dtdg.dataloader` package."""
 
 import numpy as np
 from tests.core.common_functions import *
@@ -9,11 +9,11 @@ from tests.core.common_functions import *
 def test_ogb_dataset_torch():
     clear_background()
     """Test that citation graph could support different backend."""
-    from nineturn.core.config import set_backend
-    from nineturn.core.backends import PYTORCH
+    from spiro.core.config import set_backend
+    from spiro.core.backends import PYTORCH
 
     set_backend(PYTORCH)
-    from nineturn.dtdg.dataloader import ogb_dataset, supported_ogb_datasets
+    from spiro.dtdg.dataloader import ogb_dataset, supported_ogb_datasets
 
     data_to_test = supported_ogb_datasets()[1]
     this_graph = ogb_dataset(data_to_test)
